Added yarn install step and updated package.json
A new step for yarn installation has been added to the GitHub workflow. This ensures that dependencies are installed using the frozen lockfile option, which helps maintain consistent installations across different environments. The working directory for this step is set as 'src'.

In addition, changes have been made to the package.json file in the 'src' directory. The license field has been added with a value of "MIT". The order of devDependencies and dependencies fields has been swapped, but their content remains unchanged. Lastly, the packageManager field's value has been simplified from a SHA-1 hash version to just "yarn@1.22.19".