Add hosts repo (#22)
* Updated class inheritance in PublicRepo

The PublicRepo class was incorrectly inheriting from the PrivateRepo class. This has been corrected to inherit from the correct PublicRepo class.

* Add public repo for server infrastructure

A new public repository has been added to manage the on-premises server infrastructure. This is in addition to the existing private repositories, and it will help streamline management of different aspects of our system.