Renovate config repo (#124)
* Added topics and additional required checks to repos

The code changes include the addition of a 'topics' field in the PublicRepoArgs interface. This allows for the specification of repository topics. The 'topics' field has also been added to the PublicRepo class, enabling it to be passed as an argument when creating new repositories.

In addition, more required checks have been added to the minecraftManager and renovateConfig repositories. These checks ensure that certain conditions are met before changes can be merged into these repositories.

* Add topics to renovateConfig in index.ts

Added an array of topics to the 'renovateConfig' object. This will help with categorization and searchability of the repository. The added topics include 'renovate', 'cicd', and 'bun'.